contract refactors, todo add kv.del
