+++ circuits/README.md
@@ -1,0 +1,6 @@
+# Circuits
+
+Circuit artifacts such as WASM file, prover key and verification key that are used by HollowDB are stored here.
+
+- A **prover** will use the WASM file and prover key to generate proofs.
+- A **verifier** can verify the generated proof by using the verification key.
