+++ contracts/README.md
@@ -1,0 +1,3 @@
+# Contracts
+
+Although there is only one contract implementation here right now, we might group multiple contracts under this folder with their respective names as the folder name.
