+++ contracts/hollowDB/actions/crud/get.ts
@@ -10,6 +10,7 @@ export const get: HollowDBContractFunction<HollowDBGet> = async (state, action) 
   const {key} = action.input.data;
 
   return {
-    result: await SmartWeave.kv.get<string>(key),
+    // @todo can this return anything other than string?
+    result: (await SmartWeave.kv.get(key)) as string,
   };
 };
