+++ contracts/hollowDB/actions/crud/update.ts
@@ -1,6 +1,6 @@
 import errors from '../../errors';
 import type {HollowDBContractFunction} from '../../types';
-import {verifyProof, valueToBigInt} from '../../../utils';
+import {verifyProof, valueToBigInt} from '../../utils';
 
 export type HollowDBUpdate = {
   function: 'update';
@@ -19,7 +19,7 @@ export const update: HollowDBContractFunction<HollowDBUpdate> = async (state, ac
   }
 
   // there must be a value at the key
-  const dbValue = await SmartWeave.kv.get<string>(key);
+  const dbValue = await SmartWeave.kv.get(key);
   if (dbValue === null) {
     throw errors.KeyNotExistsError;
   }
